requirements.txt
pyOpenSSL==23.3.0
python-dotenv==1.0.0